#188 - move meta extraction to MetasExtractor class
